+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Keep it human-readable, your future self will thank you!
 
 ## [Unreleased](https://github.com/ecmwf/anemoi-datasets/compare/0.5.0...HEAD)
 
+### Added
+
+- Adding the user recipe in the dataset PR #59.
+
 ### Changed
 
 - Bug fix in create/rename